refactor: Splitting into several modules
