+++ Cargo.toml
@@ -1,11 +1,13 @@
-[package]
-name = "vortex"
-version = "0.1.0"
-edition = "2021"
+[workspace]
+resolver = "2"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+members = [
+    "vortex",
+    "signaling",
+    "rtc", "api",
+]
 
-[dependencies]
+[workspace.dependencies]
 # Error Handling
 anyhow = "1.0.89"
 thiserror = "1.0.64"
@@ -28,4 +30,6 @@ tokio = { version = "1.40.0", features = ["full"] }
 log = "0.4.22"
 dashmap = "6.1.0"
 lazy_static = "1.5.0"
+dotenv = "0.15.0"
 pretty_env_logger = "0.5.0"
+env_logger = "0.11.5"
