+++ api/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "api"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+serde.workspace = true
+webrtc.workspace = true
+thiserror.workspace = true
