+++ api/src/lib.rs
@@ -1,0 +1,21 @@
+pub mod media_type;
+pub mod server_error;
+pub mod negotiation;
+pub mod remote_track;
+pub mod ice_candidate;
+
+
+pub fn add(left: u64, right: u64) -> u64 {
+    left + right
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn it_works() {
+        let result = add(2, 2);
+        assert_eq!(result, 4);
+    }
+}
