+++ signaling/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "signaling"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+anyhow.workspace = true
+futures.workspace = true
+postage.workspace = true
+log.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+tokio.workspace = true
+tokio-tungstenite.workspace = true
+webrtc.workspace = true
+thiserror.workspace = true
+
+api = { path = "../api" }
+rtc = { path = "../rtc" }
