+++ signaling/src/client.rs
@@ -5,15 +5,16 @@ use futures::{
     future::{select, Either},
     pin_mut, FutureExt, TryStreamExt,
 };
+use log::{debug, info};
 use postage::stream::Stream;
+use api::negotiation::Negotiation;
+use api::server_error::ServerError;
+use rtc::peer::Peer;
+use rtc::room::{Room, RoomEvent};
 
-use crate::rtc::{
-    peer::Peer,
-    room::{Room, RoomEvent},
-};
 
 use super::{
-    packets::{Negotiation, PacketC2S, PacketS2C, ServerError},
+    packets::{PacketC2S, PacketS2C},
     sender::{ReadWritePair, Sender},
     server::UserInformation,
 };
