+++ vortex/src/main.rs
@@ -1,24 +1,26 @@
 #[macro_use]
 extern crate log;
-#[macro_use]
 extern crate serde;
-#[macro_use]
 extern crate lazy_static;
 
 use anyhow::Result;
-
-pub mod rtc;
-pub mod signaling;
+use api::server_error::ServerError;
 
 #[tokio::main]
 async fn main() -> Result<()> {
-    pretty_env_logger::init();
+    dotenv::dotenv().ok();
+    env_logger::init_from_env(env_logger::Env::default()
+        .filter_or("RUST_LOG", "info"));
+
+    let address = dotenv::var("HTTP_HOST")
+        .unwrap_or_else(|_| "127.0.0.1:3000".to_string());
+
+    info!("Starting vortex server at {}", address);
 
     signaling::server::launch(
-        "0.0.0.0:3000",
+        address,
         Box::new(move |room_id, token| {
             Box::pin(async move {
-                use signaling::packets::ServerError;
                 if room_id != "1" {
                     return Err(ServerError::RoomNotFound.into());
                 }
